DAWeb/requirement.txt
djangorestframework
psycopg2
djangorestframework-simplejwt
requests